requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
boto3==1.34.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
numpy>=1.26.0
pandas>=2.0.0
python-dotenv==1.0.0 